Refactor DOM engines to use more scalable structure
